Added an archiving thing to let people go through posts by date
- You can now go through posts by date like any other blog.
- There's now a sidebar both in post detail views and the posts views

Note:
This whole thing needs to be worked at a bit more. The archives have
no pagination and can really clutter the browser if there's more than
like 150 posts. The sidebar is fixed and should only display months
which contain posts as well as limiting the number of years listed
to like 2 or 3, minimizing the rest to be clickable and go to a proper
archiving view with the full list.